requirements.txt
pydantic>=2.0.0
openai>=1.0.0
anthropic>=0.3.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pypdf>=3.0.0
flask>=3.0.0
flask-cors>=4.0.0
